# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Eco low watt, Balanced default, Turbo max profit, Solar matches battery %
 */
export type PowerProfile = (typeof PowerProfile)[keyof typeof PowerProfile];

export const PowerProfile = {
  eco: "eco",
  balanced: "balanced",
  turbo: "turbo",
  solar: "solar",
} as const;

export type BuddyMood = (typeof BuddyMood)[keyof typeof BuddyMood];

export const BuddyMood = {
  happy: "happy",
  hyped: "hyped",
  sleepy: "sleepy",
  sad: "sad",
  hungry: "hungry",
  sick: "sick",
  offline: "offline",
} as const;

export type MinerStatusMode =
  (typeof MinerStatusMode)[keyof typeof MinerStatusMode];

export const MinerStatusMode = {
  demo: "demo",
  real: "real",
} as const;

export interface MinerStatus {
  running: boolean;
  mode: MinerStatusMode;
  profile: PowerProfile;
  mood: BuddyMood;
  buddyName: string;
  buddyLevel: number;
  buddyXp: number;
  /** 0..100 */
  buddyHunger: number;
  /** gigahashes/sec (display value) */
  hashrateGhs: number;
  powerWatts: number;
  /** USD/day per watt */
  profitPerWatt: number;
  estDailyUsd: number;
  estMonthlyUsd: number;
  sharesAccepted: number;
  sharesRejected: number;
  bestDifficulty: number;
  workers: number;
  uptimeSec: number;
  ipAddress: string;
  overclockActive: boolean;
  overclockMultiplier: number;
  /** 0..100 risk */
  overclockHeat: number;
  overclockCooldownSec: number;
  /** solar profile battery % */
  batteryPercent: number;
  streakDays: number;
  btcPriceUsd: number;
  timestamp: string;
}

export interface OverclockBody {
  /**
   * @minimum 1
   * @maximum 100
   */
  clicks: number;
}

export interface OverclockResult {
  accepted: boolean;
  multiplier: number;
  heat: number;
  cooldownSec: number;
  reason?: string | null;
}

export type FeedBodyAction =
  (typeof FeedBodyAction)[keyof typeof FeedBodyAction];

export const FeedBodyAction = {
  feed: "feed",
  pet: "pet",
  clean: "clean",
} as const;

export interface FeedBody {
  action: FeedBodyAction;
}

export interface PowerProfileState {
  profile: PowerProfile;
  targetWatts: number;
  targetHashrateGhs: number;
  batteryPercent: number;
}

export interface SetPowerProfileBody {
  profile: PowerProfile;
  /**
   * @minimum 0
   * @maximum 100
   */
  batteryPercent?: number | null;
}

export type MinerConfigMode =
  (typeof MinerConfigMode)[keyof typeof MinerConfigMode];

export const MinerConfigMode = {
  demo: "demo",
  real: "real",
} as const;

export interface MinerConfig {
  mode: MinerConfigMode;
  buddyName: string;
  walletAddress?: string | null;
  poolUrl: string;
  workerName: string;
  /** USD per kWh */
  electricityCostKwh: number;
  pcMinerEnabled: boolean;
}

export type UpdateMinerConfigBodyMode =
  | (typeof UpdateMinerConfigBodyMode)[keyof typeof UpdateMinerConfigBodyMode]
  | null;

export const UpdateMinerConfigBodyMode = {
  demo: "demo",
  real: "real",
} as const;

export interface UpdateMinerConfigBody {
  mode?: UpdateMinerConfigBodyMode;
  buddyName?: string | null;
  walletAddress?: string | null;
  poolUrl?: string | null;
  workerName?: string | null;
  electricityCostKwh?: number | null;
  pcMinerEnabled?: boolean | null;
}

export interface TelemetryPoint {
  t: string;
  hashrateGhs: number;
  powerWatts: number;
  earningsUsd: number;
  sharesAccepted: number;
  sharesRejected: number;
  deviceId: string;
}

export interface ProfitHeatmapHour {
  /**
   * @minimum 0
   * @maximum 23
   */
  hour: number;
  /** 0..100 */
  profitScore: number;
  netUsd: number;
  avgPriceUsd: number;
  avgPowerCostUsd: number;
}

export interface ProfitHeatmap {
  hours: ProfitHeatmapHour[];
  bestHour: number;
  worstHour: number;
}

export interface Achievement {
  id: string;
  name: string;
  description: string;
  /** emoji-free sticker code */
  sticker: string;
  earnedAt?: string | null;
  /** 0..1 */
  progress: number;
}

export interface Mission {
  id: string;
  title: string;
  target: number;
  current: number;
  unit: string;
  rewardXp: number;
  completed: boolean;
}

export interface PriceQuote {
  usd: number;
  change24hPct: number;
  source: string;
  timestamp: string;
}

export type DeviceKind = (typeof DeviceKind)[keyof typeof DeviceKind];

export const DeviceKind = {
  pi: "pi",
  pc: "pc",
  esp32: "esp32",
  sim: "sim",
} as const;

export interface Device {
  id: string;
  label: string;
  kind: DeviceKind;
  online: boolean;
  hashrateGhs: number;
  powerWatts: number;
  lastSeen: string;
}

export type IngestTelemetryBodyKind =
  (typeof IngestTelemetryBodyKind)[keyof typeof IngestTelemetryBodyKind];

export const IngestTelemetryBodyKind = {
  pi: "pi",
  pc: "pc",
  esp32: "esp32",
  sim: "sim",
} as const;

export interface IngestTelemetryBody {
  deviceId: string;
  label: string;
  kind: IngestTelemetryBodyKind;
  hashrateGhs: number;
  powerWatts: number;
  sharesAccepted: number;
  sharesRejected: number;
}

export interface IngestTelemetryAck {
  ok: boolean;
  profile: PowerProfile;
  suggestedHashrateGhs: number;
}

export type AiPresetCategory =
  (typeof AiPresetCategory)[keyof typeof AiPresetCategory];

export const AiPresetCategory = {
  automation: "automation",
  optimization: "optimization",
  hardware: "hardware",
  solar: "solar",
  gamify: "gamify",
  connector: "connector",
} as const;

export interface AiPreset {
  id: string;
  label: string;
  prompt: string;
  category: AiPresetCategory;
}

export type AiChatBodyModel =
  (typeof AiChatBodyModel)[keyof typeof AiChatBodyModel];

export const AiChatBodyModel = {
  claude: "claude",
  gpt: "gpt",
} as const;

export interface AiChatBody {
  message: string;
  model: AiChatBodyModel;
}

export type GetMinerHistoryParams = {
  /**
   * @minimum 1
   * @maximum 1440
   */
  minutes?: number;
};