# lib/api-zod/src/generated/types/buddyMood.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */

export type BuddyMood = (typeof BuddyMood)[keyof typeof BuddyMood];

export const BuddyMood = {
  happy: "happy",
  hyped: "hyped",
  sleepy: "sleepy",
  sad: "sad",
  hungry: "hungry",
  sick: "sick",
  offline: "offline",
} as const;