# lib/api-zod/src/generated/types/ingestTelemetryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */
import type { IngestTelemetryBodyKind } from "./ingestTelemetryBodyKind";

export interface IngestTelemetryBody {
  deviceId: string;
  label: string;
  kind: IngestTelemetryBodyKind;
  hashrateGhs: number;
  powerWatts: number;
  sharesAccepted: number;
  sharesRejected: number;
}