# lib/api-zod/src/generated/types/minerStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */
import type { BuddyMood } from "./buddyMood";
import type { MinerStatusMode } from "./minerStatusMode";
import type { PowerProfile } from "./powerProfile";

export interface MinerStatus {
  running: boolean;
  mode: MinerStatusMode;
  profile: PowerProfile;
  mood: BuddyMood;
  buddyName: string;
  buddyLevel: number;
  buddyXp: number;
  /** 0..100 */
  buddyHunger: number;
  /** gigahashes/sec (display value) */
  hashrateGhs: number;
  powerWatts: number;
  /** USD/day per watt */
  profitPerWatt: number;
  estDailyUsd: number;
  estMonthlyUsd: number;
  sharesAccepted: number;
  sharesRejected: number;
  bestDifficulty: number;
  workers: number;
  uptimeSec: number;
  ipAddress: string;
  overclockActive: boolean;
  overclockMultiplier: number;
  /** 0..100 risk */
  overclockHeat: number;
  overclockCooldownSec: number;
  /** solar profile battery % */
  batteryPercent: number;
  streakDays: number;
  btcPriceUsd: number;
  timestamp: Date;
}