# lib/api-zod/src/generated/types/overclockBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */

export interface OverclockBody {
  /**
   * @minimum 1
   * @maximum 100
   */
  clicks: number;
}