# lib/api-zod/src/generated/types/overclockResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */

export interface OverclockResult {
  accepted: boolean;
  multiplier: number;
  heat: number;
  cooldownSec: number;
  reason?: string | null;
}