# lib/api-zod/src/generated/types/powerProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Eco low watt, Balanced default, Turbo max profit, Solar matches battery %
 */
export type PowerProfile = (typeof PowerProfile)[keyof typeof PowerProfile];

export const PowerProfile = {
  eco: "eco",
  balanced: "balanced",
  turbo: "turbo",
  solar: "solar",
} as const;