# lib/api-zod/src/generated/types/setPowerProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */
import type { PowerProfile } from "./powerProfile";

export interface SetPowerProfileBody {
  profile: PowerProfile;
  /**
   * @minimum 0
   * @maximum 100
   */
  batteryPercent?: number | null;
}