# lib/api-zod/src/generated/types/telemetryPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guama-Yotchi-Biti-Miner API
 * OpenAPI spec version: 0.1.0
 */

export interface TelemetryPoint {
  t: Date;
  hashrateGhs: number;
  powerWatts: number;
  earningsUsd: number;
  sharesAccepted: number;
  sharesRejected: number;
  deviceId: string;
}